Move RAS_CubeMapManager in separated files.
