+++ source/gameengine/Rasterizer/RAS_CubeMap.cpp
@@ -15,12 +15,7 @@
 * along with this program; if not, write to the Free Software Foundation,
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 *
-* The Original Code is Copyright (C) 2001-2002 by NaN Holding BV.
-* All rights reserved.
-*
-* The Original Code is: all of this file.
-*
-* Contributor(s): none yet.
+* Contributor(s): Ulysse Martin, Tristan Porteries.
 *
 * ***** END GPL LICENSE BLOCK *****
 */
@@ -30,80 +25,17 @@
 */
 
 #include "RAS_CubeMap.h"
-#include "RAS_MeshObject.h"
-#include "RAS_IPolygonMaterial.h"
+#include "RAS_CubeMapManager.h"
+#include "RAS_Texture.h"
 #include "RAS_IRasterizer.h"
 
 #include "GPU_texture.h"
 #include "GPU_framebuffer.h"
 #include "GPU_draw.h"
 
+#include "DNA_texture_types.h"
+
 #include "glew-mx.h"
-
-#include "BKE_image.h"
-#include "BKE_global.h"
-
-#include "KX_GameObject.h"
-
-MT_Matrix4x4 bottomFaceViewMat(
-	-1.0f, 0.0f, 0.0f, 0.0f,
-	0.0f, -1.0f, 0.0f, 0.0f,
-	0.0f, 0.0f, 1.0f, 0.0f,
-	0.0f, 0.0f, 0.0f, 1.0f);
-MT_Matrix4x4 topFaceViewMat(
-	1.0f, 0.0f, 0.0f, 0.0f,
-	0.0f, -1.0f, 0.0f, 0.0f,
-	0.0f, 0.0f, -1.0f, 0.0f,
-	0.0f, 0.0f, 0.0f, 1.0f);
-MT_Matrix4x4 rightFaceViewMat(
-	1.0f, 0.0f, 0.0f, 0.0f,
-	0.0f, 0.0f, 1.0f, 0.0f,
-	0.0f, -1.0f, 0.0f, 0.0f,
-	0.0f, 0.0f, 0.0f, 1.0f);
-MT_Matrix4x4 leftFaceViewMat(
-	1.0f, 0.0f, 0.0f, 0.0f,
-	0.0f, 0.0f, -1.0f, 0.0f,
-	0.0f, 1.0f, 0.0f, 0.0f,
-	0.0f, 0.0f, 0.0f, 1.0f);
-MT_Matrix4x4 backFaceViewMat(
-	0.0f, 0.0f, 1.0f, 0.0f,
-	0.0f, -1.0f, 0.0f, 0.0f,
-	1.0f, 0.0f, 0.0f, 0.0f,
-	0.0f, 0.0f, 0.0f, 1.0f);
-MT_Matrix4x4 frontFaceViewMat(
-	0.0f, 0.0f, -1.0f, 0.0f,
-	0.0f, -1.0f, 0.0f, 0.0f,
-	-1.0f, 0.0f, 0.0f, 0.0f,
-	0.0f, 0.0f, 0.0f, 1.0f);
-
-static MT_Matrix3x3 bottomCamOri(
-	-1.0f, 0.0f, 0.0f,
-	0.0f, -1.0f, 0.0f,
-	0.0f, 0.0f, 1.0f);
-static MT_Matrix3x3 topCamOri(
-	1.0f, 0.0f, 0.0f,
-	0.0f, -1.0f, 0.0f,
-	0.0f, 0.0f, -1.0f);
-static MT_Matrix3x3 rightCamOri(
-	1.0f, 0.0f, 0.0f,
-	0.0f, 0.0f, 1.0f,
-	0.0f, -1.0f, 0.0f);
-static MT_Matrix3x3 leftCamOri(
-	1.0f, 0.0f, 0.0f,
-	0.0f, 0.0f, -1.0f,
-	0.0f, 1.0f, 0.0f);
-static MT_Matrix3x3 backCamOri(
-	0.0f, 0.0f, 1.0f,
-	0.0f, -1.0f, 0.0f,
-	1.0f, 0.0f, 0.0f);
-static MT_Matrix3x3 frontCamOri(
-	0.0f, 0.0f, -1.0f,
-	0.0f, -1.0f, 0.0f,
-	-1.0f, 0.0f, 0.0f);
-
-MT_Matrix4x4 RAS_CubeMapManager::facesViewMat[6] = { topFaceViewMat, bottomFaceViewMat, frontFaceViewMat, backFaceViewMat, rightFaceViewMat, leftFaceViewMat };
-MT_Matrix3x3 RAS_CubeMapManager::camOri[6] = { topCamOri, bottomCamOri, frontCamOri, backCamOri, rightCamOri, leftCamOri };
-MT_Matrix3x3 RAS_CubeMapManager::camOri2[6] = { topCamOri, bottomCamOri, frontCamOri, backCamOri, leftCamOri, rightCamOri };
 
 RAS_CubeMap::RAS_CubeMap(void *clientobj, RAS_Texture *texture, RAS_IRasterizer *rasty)
 	:m_texture(texture),
@@ -189,37 +121,4 @@ void RAS_CubeMap::UnbindFace()
 {
 	GPU_framebuffer_texture_detach(m_cubeMapTexture);
 	GPU_framebuffer_texture_unbind(m_fbo, m_cubeMapTexture);
-}
-
-RAS_CubeMapManager::RAS_CubeMapManager()
-{
-}
-
-RAS_CubeMapManager::~RAS_CubeMapManager()
-{
-	for (std::vector<RAS_CubeMap *>::iterator it = m_cubeMaps.begin(), end = m_cubeMaps.end(); it != end; ++it) {
-		delete *it;
-	}
-}
-
-void RAS_CubeMapManager::AddCubeMap(RAS_CubeMap *cubeMap)
-{
-	m_cubeMaps.push_back(cubeMap);
-}
-
-void RAS_CubeMapManager::RemoveCubeMap(void *clientobj)
-{
-	for (std::vector<RAS_CubeMap *>::iterator it = m_cubeMaps.begin(), end = m_cubeMaps.end(); it != end; ++it) {
-		RAS_CubeMap *cubeMap = *it;
-		if (cubeMap->GetClientObject() == clientobj) {
-			delete cubeMap;
-			m_cubeMaps.erase(it);
-			break;
-		}
-	}
-}
-
-void RAS_CubeMapManager::RestoreFrameBuffer()
-{
-	GPU_framebuffer_restore();
 }
